backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
boto3>=1.34.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-multipart>=0.0.9
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
moto[dynamodb,s3]>=5.0.0
